submit order to stripe

+++ client/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -82,11 +82,28 @@ export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
   submitOrder() {
     const basket = this.basketService.getCurrentBasketValue();
     const orderToCreate = this.getOrderToCreate(basket);
-    this.checkoutService.creatOrder(orderToCreate).subscribe((order: IOrder) => {
-      this.toastr.success('Order created successfully');
-      this.basketService.deleteLocalBasket(basket.id);
-      const navigationExtras: NavigationExtras = { state: order };
-      this.router.navigate(['checkout/success'], navigationExtras);
+    this.checkoutService.createOrder(orderToCreate).subscribe((order: IOrder) => {
+    this.toastr.success('Order created successfully'); 
+    this.stripe.confirmCardPayment(basket.clientSecret, {
+        payment_method: {
+          card: this.cardNumber,
+          billing_details: {
+            name: this.checkoutForm.get('paymentForm').get('nameOnCard').value,
+          },
+        },
+      })
+      .then(function (result) {
+        console.log(result)
+        console.log(basket.id, 'basket id')
+        if (result.paymentIntent) {
+          this.basketService.deleteLocalBasket(basket.id);
+          const navigationExtras: NavigationExtras = { state: order };
+          this.router.navigate(['checkout/success'], navigationExtras);
+        } else {
+          this.toastr.error('Payment failed')
+        }
+      });
+     
     }, error => {
       this.toastr.error(error.message);
       console.log(error);
